+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/pedigree/PedigreeDiseaseCompatibilityDecorator.java
@@ -1,5 +1,10 @@
 package de.charite.compbio.jannovar.pedigree;
 
+import de.charite.compbio.jannovar.pedigree.compatibilitychecker.ad.CompatibilityCheckerAutosomalDominant;
+import de.charite.compbio.jannovar.pedigree.compatibilitychecker.ar.CompatibilityCheckerAutosomalRecessive;
+import de.charite.compbio.jannovar.pedigree.compatibilitychecker.xd.CompatibilityCheckerXDominant;
+import de.charite.compbio.jannovar.pedigree.compatibilitychecker.xr.CompatibilityCheckerXRecessive;
+
 // TODO(holtgrewe): Add functionality to compute marker for non-compa
 
 /**