+++ jannovar-core/src/test/java/de/charite/compbio/jannovar/pedigree/CompatibilityCheckerXRecessiveStupidSmallFemaleTest.java
@@ -10,6 +10,9 @@ public class CompatibilityCheckerXRecessiveStupidSmallFemaleTest extends Compati
 
 	@Before
 	public void setUp() throws PedParseException {
+		// STUPID PEDIGREE! The male must be affected. Otherwise it can only be
+		// a de-novo second second hit. But we do not want to cover special
+		// cases. So this case should always return no variant!
 		ImmutableList.Builder<PedPerson> individuals = new ImmutableList.Builder<PedPerson>();
 		individuals.add(new PedPerson("ped", "I.1", "0", "0", Sex.MALE, Disease.UNAFFECTED)); // father
 		individuals.add(new PedPerson("ped", "I.2", "0", "0", Sex.FEMALE, Disease.UNAFFECTED)); // mother